Flashrom BusPirate_SPI LowVoltage flash support continued
NOTE 201702:
I am commiting these changes to git after almost 1.5 years of doing these
changes, so my comments could be vague and or slightly incorrect. As of
now I am just logically guessing stuff and writing these commit comment,
instead of going thro the documents of things (buspirate, flash chips,
...) in detail before commiting.

* Reduced the spispeed by default

Better defaults for LowVoltage (1.8v or lower) Flash chips, which use their
own supply voltage (so no need for BusPirate to provide power). Also in this
case as the SPI lines require to work wrt the low io voltage of these chips,
the buspirate internal referencing of spi to 3.3v is disabled and it is put
into HiZ mode, so that a external (wrt the uC) pullup can be used to provide
the logic 1 reference as required. Inturn as BusPirate already provides
provision for having pullups on these uC IO lines in its HW, the same is
enabled by default, so that there is no need for mounting/connecting pullup
resistors externally between the BusPirate and the Flash chip. However this
also means that one is required to provide the io reference voltage to the
Vpu/Vextern pin of buspirate.

* Have enabled pullup by default now
  So if off is passed wrt pullups state, it is explicitly disabled now.
  NOTE: One is required to provide a external reference voltage to Vpu pin
  of buspirate.

* Have added a new option/param hiz and inturn corresponding variable outhiz.

  * By default outhiz is enabled

  * If External Power Mode is selected, then also outhiz is enabled

  * user can explicitly control thro the hiz parameter

  * instead of my previous powerext == 1 check to put the spi related pins
    to HiZ mode instead of 3.3v mode, now this outhiz is checked.

* Added few debug print messages wrt spi read.